arch: Move everything inside devito/arch
