+++ devito/__init__.py
@@ -16,7 +16,7 @@ from devito.types.relational import *  # noqa
 from devito.types.tensor import *  # noqa
 
 # Imports required to initialize Devito
-from devito.archinfo import platform_registry
+from devito.arch import platform_registry
 from devito.backends import backends_registry, init_backend
 from devito.compiler import compiler_registry
 from devito.logger import error, warning, info, logger_registry, set_log_level  # noqa
